Added cluster failover testing docs
This doc outlines what would happen when the central cluster goes down.

This test simulates a scenario in which the central cluster goes down. It outlines what happens to a topic's data, leader, ISRs and how a cluster would handle such a failover scenario. Later in the test, the central failed cluster is recovered to see if the old state is maintained by the strected kafka clusters.